KSM 2023/10/24 add post-service(Blog entity)
블로그 엔티티에서 User의 정보를 불러와서 외래키로 설정하려 했지만 오류가 생김

+++ member-service/src/main/java/com/example/memberservice/entity/Authentication.java
@@ -10,7 +10,7 @@ public class Authentication extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "AuthenticationNo")
-    private long no;
+    private int no;
 
     @OneToOne
     @JoinColumn(name = "UserNo", referencedColumnName = "UserNo")
